ingest/requirements.txt
httpx==0.27.2 #used for requests, its light
python-dotenv==1.0.1  # for loading .env in the container
minio==7.2.15         # S3-compatible client for MinIO
requests==2.32.3      # fallback HTTP client for REST calls (optional)

sec-api==1.0.32